share one oauth token stub across the tests
