+++ FEDERATION.md
@@ -229,7 +229,7 @@ When the ActivityPub API option is enabled, the plugin exposes OAuth 2.0 endpoin
 
 **Supported standards:**
 
-- [SWICG ActivityPub API: Basic Profile](https://swicg.github.io/activitypub-api/basicprofile) - C2S baseline. The token response includes `activitypub_actor_id` alongside the IndieAuth `me` URI. Two generations of Basic Profile scope identifiers are accepted and advertised in `scopes_supported`: the `activitypub:read:*` / `activitypub:write:*` aliases the draft used before 2026-08-04, and the `https://swicg.github.io/activitypub-api/scopes#…` identifiers that replaced them. Each is mapped to the scope the plugin grants, so the per-collection read identifiers resolve to `read` and the per-action write identifiers to `write`. There is no per-activity access control, so `follow` and `updateprofile` resolve to `write` like every other write identifier. Three identifiers are not offered: `readown` and `reactown` describe data on the client's own server, and `uploadfiles` needs a MediaUpload endpoint the plugin does not implement.
+- [SWICG ActivityPub API: Basic Profile](https://swicg.github.io/activitypub-api/basicprofile) - C2S baseline. The token response includes `activitypub_actor_id` alongside the IndieAuth `me` URI. Two generations of Basic Profile scope identifiers are accepted and advertised in `scopes_supported`: the `activitypub:read:*` / `activitypub:write:*` aliases the draft used before 2026-08-04, and the `https://swicg.github.io/activitypub-api/scopes#…` identifiers that replaced them. Each is mapped to the scope the plugin grants, so the per-collection read identifiers resolve to `read` and the per-action write identifiers to `write`. There is no per-activity access control, so `follow` and `updateprofile` resolve to `write` like every other write identifier. Seven identifiers are not offered: `readown` and `reactown` describe data on the client's own server, `uploadfiles` needs a MediaUpload endpoint the plugin does not implement, and the four `address*` identifiers restrict who an activity may be addressed to rather than granting anything.
 - [RFC 8252](https://datatracker.ietf.org/doc/html/rfc8252) - OAuth 2.0 for Native Apps. Loopback redirect URIs (`http://127.0.0.1:{port}` and `http://[::1]:{port}`) are accepted with port flexibility per §7.3/§8.3. `localhost` is also accepted for compatibility; §8.3 marks this "NOT RECOMMENDED" but it remains common practice.
 - [RFC 7591](https://datatracker.ietf.org/doc/html/rfc7591) - Dynamic Client Registration.
 - [RFC 7636](https://datatracker.ietf.org/doc/html/rfc7636) - PKCE. Required by default for public clients; only `S256` is accepted.
