+++ includes/oauth/class-scope.php
@@ -119,13 +119,14 @@ class Scope {
 	 *
 	 * Each identifier maps to a scope the plugin grants, which is coarser than the spec intends:
 	 * the per-collection read identifiers all resolve to `read`, and the per-action write
-	 * identifiers, `follow` and `updateprofile` included, to `write`. Three identifiers are
+	 * identifiers, `follow` and `updateprofile` included, to `write`. Seven identifiers are
 	 * deliberately absent, because the plugin has nothing to grant for them: `readown` and
-	 * `reactown` describe data on the client's own server rather than this one, `uploadfiles`
-	 * needs a MediaUpload endpoint the plugin does not implement, and the four `address*`
-	 * identifiers narrow who an activity may be addressed to, which is a restriction on a write
-	 * rather than a permission of its own. Resolving those to `write` would answer a request to
-	 * be limited by handing over the unlimited version.
+	 * `reactown` describe data on the client's own server rather than this one; `uploadfiles`
+	 * needs a MediaUpload endpoint the plugin does not implement; and `addressall`,
+	 * `addresspublic`, `addressactor` and `addressfollowers` narrow who an activity may be
+	 * addressed to, which is a restriction on a write rather than a permission of its own.
+	 * Resolving those to `write` would answer a request to be limited by handing over the
+	 * unlimited version.
 	 *
 	 * @since unreleased
 	 */
